+++ bcbio/rnaseq/featureCounts.py
@@ -43,6 +43,12 @@ def count(data):
     cmd = ("{featureCounts} -a {gtf_file} -o {tx_count_file} -s {strand_flag} "
            "{paired_flag} {filtered_bam}")
 
+    resources = config_utils.get_resources("featureCounts", data["config"])
+    if resources:
+        options = resources.get("options")
+        if options:
+            cmd += " %s" % " ".join([str(x) for x in options])
+
     message = ("Count reads in {tx_count_file} mapping to {gtf_file} using "
                "featureCounts")
     with file_transaction(data, [count_file, summary_file]) as tx_files: